Work with Data JPA

+++ spring-start-aspect/src/main/java/net/aspect/education/database/dto/CompanyReadDto.java
@@ -1,4 +1,4 @@
 package net.aspect.education.database.dto;
 
-public record CompanyReadDto(Long id) {
+public record CompanyReadDto(Integer id) {
 }
